requirements.txt
aiopg==1.4.0
alembic==1.12.1
fastapi==0.104.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
uvicorn==0.23.2
